choieunbi 님이 변경한 내용

+++ src/main/java/com/prgrms/prolog/domain/post/service/PostServiceImpl.java
@@ -12,9 +12,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.prgrms.prolog.domain.post.dto.PostRequest.CreateRequest;
-import com.prgrms.prolog.domain.post.dto.PostRequest.UpdateRequest;
-import com.prgrms.prolog.domain.post.dto.PostResponse;
+import com.prgrms.prolog.domain.post.dto.PostDto.CreatePostRequest;
+import com.prgrms.prolog.domain.post.dto.PostDto.SinglePostResponse;
+import com.prgrms.prolog.domain.post.dto.PostDto.UpdatePostRequest;
+import com.prgrms.prolog.domain.post.exception.IllegalAccessPostException;
 import com.prgrms.prolog.domain.post.model.Post;
 import com.prgrms.prolog.domain.post.repository.PostRepository;
 import com.prgrms.prolog.domain.posttag.model.PostTag;
@@ -36,12 +37,9 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class PostServiceImpl implements PostService {
 
-	private static final String POST_NOT_EXIST_MESSAGE = "존재하지 않는 게시물입니다.";
-	private static final String USER_NOT_EXIST_MESSAGE = "존재하지 않는 사용자입니다.";
-
-	private final SeriesRepository seriesRepository;
-	private final PostRepository postRepository;
 	private final UserRepository userRepository;
+	private final PostRepository postRepository;
+	private final SeriesRepository seriesRepository;
 	private final RootTagRepository rootTagRepository;
 	private final PostTagRepository postTagRepository;
 	private final UserTagRepository userTagRepository;
@@ -49,16 +47,16 @@ public class PostServiceImpl implements PostService {
 
 	@Override
 	@Transactional
-	public Long create(CreateRequest createRequest, Long userId) {
+	public Long createPost(CreatePostRequest createRequest, Long userId) {
 		User findUser = userRepository.joinUserTagFindByUserId(userId);
-		Post createdPost = CreateRequest.toEntity(createRequest, findUser);
+		Post createdPost = CreatePostRequest.from(createRequest, findUser);
 		Post savedPost = postRepository.save(createdPost);
 		updateNewPostAndUserIfTagExists(createRequest.tagText(), savedPost, findUser);
 		registerSeries(createRequest, savedPost, findUser);
 		return savedPost.getId();
 	}
 
-	private void registerSeries(CreateRequest request, Post post, User owner) {
+	private void registerSeries(CreatePostRequest request, Post post, User owner) {
 		String seriesTitle = request.seriesTitle();
 		if (seriesTitle == null || seriesTitle.isBlank()) {
 			seriesTitle = "시리즈 없음";
@@ -77,47 +75,56 @@ public class PostServiceImpl implements PostService {
 	}
 
 	@Override
-	public PostResponse findById(Long postId) {
+	public SinglePostResponse getSinglePost(Long userId, Long postId) {
 		Post post = postRepository.joinCommentFindByPostId(postId)
 			.orElseThrow(() -> new IllegalArgumentException("exception.post.notExists"));
+		if (!post.isOpenStatus() && !post.getUser().checkSameUserId(userId)) { // 비공개 and 다른 사용자인 경우
+			throw new IllegalAccessPostException("exception.post.not.access");
+		}
 		Set<PostTag> findPostTags = postTagRepository.joinRootTagFindByPostId(postId);
-		post.addPostTagsFrom(findPostTags);
-		return PostResponse.from(post);
+		post.addPostTags(findPostTags);
+		return SinglePostResponse.from(post);
 	}
 
 	@Override
-	public Page<PostResponse> findAll(Pageable pageable) {
+	public Page<SinglePostResponse> getAllPost(Pageable pageable) {
 		return postRepository.findAll(pageable)
-			.map(PostResponse::from);
+			.map(SinglePostResponse::from);
 	}
 
 	@Override
 	@Transactional
-	public PostResponse update(UpdateRequest update, Long userId, Long postId) {
+	public SinglePostResponse updatePost(UpdatePostRequest updatePostRequest, Long userId, Long postId) {
 		Post findPost = postRepository.findById(postId)
-			.orElseThrow(() -> new IllegalArgumentException(POST_NOT_EXIST_MESSAGE));
+			.orElseThrow(() -> new IllegalArgumentException("exception.post.notExists"));
 
 		if (!findPost.getUser().checkSameUserId(userId)) {
 			throw new IllegalArgumentException("exception.post.not.owner");
 		}
 
-		findPost.changePost(update);
-		updateIfTagChanged(update.tagText(), findPost);
+		findPost.changePost(updatePostRequest);
+		updateIfTagChanged(updatePostRequest.tagText(), findPost);
 
 		Set<PostTag> findPostTags = postTagRepository.joinRootTagFindByPostId(findPost.getId());
-		findPost.addPostTagsFrom(findPostTags);
-		return PostResponse.from(findPost);
+		findPost.addPostTags(findPostTags);
+		return SinglePostResponse.from(findPost);
 	}
 
 	@Override
 	@Transactional
-	public void delete(Long postId) {
+	public void deletePost(Long userId, Long postId) {
 		Post findPost = postRepository.findById(postId)
-			.orElseThrow(() -> new IllegalArgumentException(POST_NOT_EXIST_MESSAGE));
+			.orElseThrow(() -> new IllegalArgumentException("exception.post.notExists"));
+
+		if (!findPost.getUser().checkSameUserId(userId)) {
+			throw new IllegalArgumentException("exception.post.not.owner");
+		}
+
 		Set<RootTag> findRootTags = postTagRepository.joinRootTagFindByPostId(findPost.getId())
 			.stream()
 			.map(PostTag::getRootTag)
 			.collect(Collectors.toSet());
+
 		removeOrDecreaseUserTags(findPost.getUser(), findRootTags);
 		postTagRepository.deleteByPostId(postId);
 		postRepository.delete(findPost);
